fix: drill down recent meetings as a real section filter and split sidebar UI
Reviewer feedback on PR #758:

- "Recent meeting notes from last 7 days" was rendered/counted but did not
  filter. The drill-down now sends `section=recent_meetings`, the backend
  scopes the queue to items linked to a meeting transcript or summary
  artifact created in the last 7 days, and the row toggles `is-active`.
- internal/web/static/app-item-sidebar-ui.ts exceeded the 1000-line module
  cap. The compact sidebar primary header, secondary expandable section,
  and queue tabs are extracted into app-item-sidebar-secondary.ts; the
  ui module drops to ~735 lines.
